backend/requirements/common.txt
django~=3.0.0
pillow
djangorestframework-jwt
djangorestframework
django-cors-headers
django-pydenticon
